+++ packages/contracts/CHANGELOG.md
@@ -18,3 +18,4 @@ and this project adheres to the [Aragon OSx Plugin Versioning Convention](https:
 - Hard-coded the `bytes32 internal constant EXECUTE_PERMISSION_ID` constant in `TokenVotingSetup` until it is available in `PermissionLib`.
 - Changed the `prepareInstallation` function in `TokenVotingSetup` to not unnecessarily grant the `UPGRADE_PLUGIN_PERMISSION_ID` permission to the installing DAO.
 - Changed the `prepareUpdate` function in `TokenVotingSetup` to revoke the previously unnecessarily granted `UPGRADE_PLUGIN_PERMISSION_ID` permission from the installing DAO.
+- Changed the `prepareUninstallation` function in `TokenVotingSetup` to not revoke the `MINT_PERMISSION_ID` to the DAO when the plugin is uninstalled.